requirements.txt
fastapi
uvicorn
slack-sdk
google-api-python-client
sqlalchemy
alembic
python-dotenv
pydantic
httpx
python-multipart
requests
